platform/sbin/apps/fastapi/requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
bcrypt==3.2.0
bmc==0.0.3
boto3==1.18.54
botocore==1.21.54
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
docutils==0.14
ecdsa==0.17.0
fastapi==0.68.1
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.2
importlib-metadata==4.8.1
Jinja2==3.0.2
jmespath==0.10.0
lockfile==0.11.0
MarkupSafe==2.0.1
minio==7.1.0
passlib==1.7.4
pika==1.2.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
pyOpenSSL==21.0.0
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
simplejson==3.2.0
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
zipp==3.5.0
psycopg2==2.9.3
kubernetes==25.3.0
pandas==1.5.3
starlette-validation-uploadfile==0.1.1
